refactor: Organize variables
